+++ lua/which-key/view.lua
@@ -52,17 +52,13 @@ function M.show()
 
   local opts = {
     relative = "editor",
-    width = vim.o.columns
-      - margins[2]
-      - margins[4]
-      - (vim.fn.has("nvim-0.6") == 0 and config.options.window.border ~= "none" and 2 or 0),
+    width = vim.o.columns - margins[2] - margins[4],
     height = config.options.layout.height.min,
     focusable = false,
     anchor = "SW",
     border = config.options.window.border,
     row = vim.o.lines
       - margins[3]
-      - (vim.fn.has("nvim-0.6") == 0 and config.options.window.border ~= "none" and 2 or 0)
       + ((vim.o.laststatus == 0 or vim.o.laststatus == 1 and #wins == 1) and 1 or 0)
       - vim.o.cmdheight,
     col = margins[4],
@@ -76,18 +72,13 @@ function M.show()
   end
   M.buf = vim.api.nvim_create_buf(false, true)
   M.win = vim.api.nvim_open_win(M.buf, false, opts)
-  vim.api.nvim_buf_set_option(M.buf, "filetype", "WhichKey")
-  vim.api.nvim_buf_set_option(M.buf, "buftype", "nofile")
-  vim.api.nvim_buf_set_option(M.buf, "bufhidden", "wipe")
-  vim.api.nvim_buf_set_option(M.buf, "modifiable", true)
-
-  local winhl = "NormalFloat:WhichKeyFloat"
-  if vim.fn.hlexists("FloatBorder") == 1 then
-    winhl = winhl .. ",FloatBorder:WhichKeyBorder"
-  end
-  vim.api.nvim_win_set_option(M.win, "winhighlight", winhl)
-  vim.api.nvim_win_set_option(M.win, "foldmethod", "manual")
-  vim.api.nvim_win_set_option(M.win, "winblend", config.options.window.winblend)
+  vim.bo[M.buf].filetype = "WhichKey"
+  vim.bo[M.buf].buftype = "nofile"
+  vim.bo[M.buf].bufhidden = "wipe"
+  vim.bo[M.buf].modifiable = true
+  vim.wo[M.win].winhighlight = "NormalFloat:WhichKeyFloat,FloatBorder:WhichKeyBorder"
+  vim.wo[M.win].foldmethod = "manual"
+  vim.wo[M.win].winblend = config.options.window.winblend
 end
 
 function M.read_pending()
@@ -204,7 +195,7 @@ function M.execute(prefix_i, mode, buf)
   end
 
   -- make sure we remove all WK hooks before executing the sequence
-  -- this is to make existing keybindongs work and prevent recursion
+  -- this is to make existing keybindings work and prevent recursion
   unhook(Keys.get_tree(mode).tree:path(prefix_i))
   if buf then
     unhook(Keys.get_tree(mode, buf).tree:path(prefix_i), buf)
@@ -256,18 +247,22 @@ function M.open(keys, opts)
 end
 
 function M.is_enabled(buf)
-  local buftype = vim.api.nvim_buf_get_option(buf, "buftype")
+  local buftype = vim.bo[buf].buftype
   for _, bt in ipairs(config.options.disable.buftypes) do
     if bt == buftype then
       return false
     end
   end
 
-  local filetype = vim.api.nvim_buf_get_option(buf, "filetype")
+  local filetype = vim.bo[buf].filetype
   for _, bt in ipairs(config.options.disable.filetypes) do
     if bt == filetype then
       return false
     end
+  end
+
+  if vim.fn.getcmdwintype() ~= "" then
+    return false
   end
 
   return true
@@ -333,6 +328,7 @@ end
 
 ---@param text Text
 function M.render(text)
+  local view = vim.api.nvim_win_call(M.win, vim.fn.winsaveview)
   vim.api.nvim_buf_set_lines(M.buf, 0, -1, false, text.lines)
   local height = #text.lines
   if height > config.options.layout.height.max then
@@ -345,6 +341,9 @@ function M.render(text)
   for _, data in ipairs(text.hl) do
     highlight(M.buf, config.namespace, data.group, data.line, data.from, data.to)
   end
+  vim.api.nvim_win_call(M.win, function()
+    vim.fn.winrestview(view)
+  end)
 end
 
 return M